fixed the float amount issue

+++ bill_Cal2/Group_bill_format.cpp
@@ -6,7 +6,7 @@
 
 
 
-Group_bill_format:: Group_bill_format(unsigned int amount, std::string * namelist, unsigned int noOfPeople,  bool isOwenerInside )
+Group_bill_format:: Group_bill_format(float amount, std::string * namelist, unsigned int noOfPeople,  bool isOwenerInside )
 {
 	if (noOfPeople > GROUP_BILL_FORMAT_MAX_PEOPLE)
 		std::cerr << "excced the max stacks in the calss group_bill_format!!\n" <<std::endl;
